requirements.txt
PyMySQL>=1.0.2
python-dotenv>=1.0.0
alembic>=1.13.0
SQLAlchemy>=2.0.0
cryptography>=46.0.0
